requirements.txt
streamlit
pandas
azure.storage.blob
python-dotenv
openpyxl
